Big changes today mate. Today I got the database portion syncing up the players names and their scores. What it does is it populates the names into an array of names after the viewDidLoad on the GameViewController. This way it ensures that the names have been added into the array. When the user hits “play again” in the GameViewController the function then increments the values in the playerScores array in the PlayerData class. Maybe I can get rid of the playerXWins and playerOWins variables since I am using the array, but I am not sure.
These two arrays are compiled together later into an object, and then that object is sent to the database and broken down into 2 key value pairs. Right now it overwrites whatever was there previously, and I need to figure out how to continually check for duplication and then append onto the database.